+++ Example/Objective-C/Coiney Test (Objective-C)/CTViewController.h
@@ -1,7 +1,7 @@
 #import <UIKit/UIKit.h>
 
 @interface CTViewController : UIViewController
-@property(weak, nonatomic) IBOutlet UITextField *productNameField, *productPriceField;
+@property(weak, nonatomic) IBOutlet UITextField *memoTextField, *priceTextField;
 
 - (IBAction)makePayment:(id)aSender;
 - (IBAction)deauthenticate:(id)aSender;